fix: move SanitizePrefix into the storage package
